Describe a linear model hypothesis to be tested

A Hypothesis can be any linear combination of coefficients, compared to zero. Specify it as a character vector that can be parsed to yield the desired equalities ala makeContrasts. A CoefficientHypothesis is a hypothesis for which terms are singly or jointly tested to be zero (generally the case in a t-test or F-test), by dropping coefficients from the model.

Hypothesis(hypothesis, terms)

Arguments

hypothesis

a character vector specifying a hypothesis, following makeContrasts, or a character vector naming coefficients to be dropped.

terms

an optional character vector giving the terms (column names from the model.matrix) out of which the contrasts will be contrasted. If missing then most functions will attempt to fill this in for you at run time.

Value

a Hypothesis with a "transformed" component

See also

zlm waldTest lrTest

Examples

h <- Hypothesis('Stim.ConditionUnstim', c('(Intercept)', 'Stim.ConditionUnstim'))

#> Warning: Some levels contain symbols.  Be careful to escape these names with backticks ('`') when specifying contrasts.

h@contrastMatrix

#>                       Contrasts
#> Levels                 Stim.ConditionUnstim
#>   (Intercept)                             0
#>   Stim.ConditionUnstim                    1