Wrapper for lmer/glmer

A horrendous hack is employed in order to do arbitrary likelihood ratio tests: the model matrix is built, the names possibly mangled, then fed in as a symbolic formula to glmer/lmer. This is necessary because there is no (easy) way to specify an arbitrary fixed-effect model matrix in glmer.

# S4 method for LMERlike
update(object, formula., design, ...)

# S4 method for LMERlike
vcov(object, which, ...)

# S4 method for LMERlike
coef(object, which, singular = TRUE, ...)

# S4 method for LMERlike
logLik(object)

Arguments

object
LMERlike

formula.
formula

design
something coercible to a data.frame

...
In the case of vcov, ignored. In the case of update, passed to model.matrix.

which
character, one of 'C', 'D'.

singular
logical. Should NA coefficients be returned?

Value

see the section "Methods (by generic)"

Methods (by generic)

update: update the formula or design matrix

vcov: return the variance/covariance of component which

coef: return the coefficients. The horrendous hack is attempted to be undone.

logLik: return the log-likelihood

Slots

pseudoMM

part of this horrendous hack.

strictConvergence

logical (default: TRUE) return results even when the optimizer or *lmer complains about convergence

optimMsg

character record warnings from lme. NA_character_ means no warnings.