Linear Model-like Class

Wrapper around modeling function to make them behave enough alike that Wald tests and Likelihood ratio are easy to do. To implement a new type of zero-inflated model, extend this class. Depending on how different the method is, you will definitely need to override the fit method, and possibly the model.matrix, model.matrix<-, update, coef, vcov, and logLik methods.

# S4 method for LMlike
summary(object)

# S4 method for LMlike
update(object, formula., design, ...)

# S4 method for LMlike,CoefficientHypothesis
waldTest(object, hypothesis)

# S4 method for LMlike,matrix
waldTest(object, hypothesis)

# S4 method for LMlike,character
lrTest(object, hypothesis)

# S4 method for LMlike,CoefficientHypothesis
lrTest(object, hypothesis)

# S4 method for LMlike,Hypothesis
lrTest(object, hypothesis)

# S4 method for LMlike,matrix
lrTest(object, hypothesis)

# S4 method for GLMlike
logLik(object)

Arguments

object

LMlike

formula.

formula

design

something coercible to a data.frame

...

passed to model.matrix

hypothesis

one of a CoefficientHypothesis, Hypothesis or contrast matrix.

Value

see section "Methods (by generic)"

Methods (by generic)

summary: Print a summary of the coefficients in each component.

update: update the formula or design from which the model.matrix is constructed

waldTest: Wald test dropping single term specified by CoefficientHypothesis hypothesis

waldTest: Wald test of contrast specified by contrast matrix hypothesis

lrTest: Likelihood ratio test dropping entire term specified by character hypothesis naming a term in the symbolic formula.

lrTest: Likelihood ratio test dropping single term specified by CoefficientHypothesis hypothesis

lrTest: Likelihood ratio test dropping single term specified by Hypothesis hypothesis

lrTest: Likelihood ratio test dropping single term specified by contrast matrix hypothesis

logLik: return the log-likelihood of a fitted model

Slots

design

a data.frame from which variables are taken for the right hand side of the regression

fitC

The continuous fit

fitD

The discrete fit

response

The left hand side of the regression

fitted

A logical with components "C" and "D", TRUE if the respective component has converged

formula

A formula for the regression

fitArgsC

fitArgsD

Both lists giving arguments that will be passed to the fitter (such as convergence criteria or case weights)

See also

coef

lrTest

waldTest

vcov

logLik