An S4 class to hold the output of a call to zlm

This holds output from a call to zlm. Many methods are defined to operate on it. See below.

# S4 method for ZlmFit,CoefficientHypothesis
lrTest(object, hypothesis)

# S4 method for ZlmFit,Hypothesis
lrTest(object, hypothesis)

# S4 method for ZlmFit,matrix
lrTest(object, hypothesis)

# S4 method for ZlmFit,CoefficientHypothesis
waldTest(object, hypothesis)

# S4 method for ZlmFit,Hypothesis
waldTest(object, hypothesis)

# S4 method for ZlmFit
coef(object, which, ...)

# S4 method for ZlmFit
vcov(object, which, ...)

# S4 method for ZlmFit
se.coef(object, which, ...)

Arguments

object
ZlmFit

hypothesis
call to Hypothesis or CoefficientHypothesis or a matrix giving such contrasts.

which
character vector, one of "C" (continuous) or "D" (discrete) specifying which component should be returned

...
ignored

Value

see "Methods (by generic)"

Methods (by generic)

lrTest: Returns an array with likelihood-ratio tests on contrasts defined using CoefficientHypothesis().

lrTest: Returns an array with likelihood-ratio tests specified by Hypothesis, which is a Hypothesis.

lrTest: Returns an array with likelihood-ratio tests specified by Hypothesis, which is a contrast matrix.

waldTest: Returns an array with Wald Tests on contrasts defined using CoefficientHypothesis().

waldTest: Returns an array with Wald Tests on contrasts defined in Hypothesis()

coef: Returns the matrix of coefficients for component which.

vcov: Returns an array of variance/covariance matrices for component which.

se.coef: Returns a matrix of standard error estimates for coefficients on component which.

Slots

coefC

matrix of continuous coefficients

coefD

matrix of discrete coefficients

vcovC

array of variance/covariance matrices for coefficients

vcovD

array of variance/covariance matrices for coefficients

LMlike

the LmWrapper object used

sca

the SingleCellAssay object used

deviance

matrix of deviances

loglik

matrix of loglikelihoods

df.null

matrix of null (intercept only) degrees of freedom

df.resid

matrix of residual DOF

dispersion

matrix of dispersions (after shrinkage)

dispersionNoShrink

matrix of dispersion (before shrinkage)

priorDOF

shrinkage weight in terms of number of psuedo-obs

priorVar

shrinkage target

converged

output that may optionally be set by the underlying modeling function

hookOut

a list of length ngenes containing output from a hook function, if zlm was called with one

See also

Examples

data(vbetaFA)
zlmVbeta <- zlm(~ Stim.Condition+Population, subset(vbetaFA, ncells==1)[1:10,])

#> 
#> Done!

#Coefficients and standard errors
coef(zlmVbeta, 'D')

#>        (Intercept) Stim.ConditionUnstim PopulationCD154-VbetaUnresponsive
#> B3GAT1  -5.4302447           1.00685352                        0.90199793
#> BAX     -0.4990626           0.02145208                       -0.89154721
#> BCL2    -3.9407287          -2.20065589                       -0.37551513
#> CCL2    -3.3547797           0.42749817                       -0.29365929
#> CCL3    -2.6590780          -1.92367996                       -0.28467650
#> CCL4    -3.3030403          -0.99943502                       -0.93080019
#> CCL5    -1.8947363           0.11842660                        0.23701817
#> CCR2    -3.7005996           1.30851899                        0.02582989
#> CCR4    -0.3695476          -0.21384188                       -0.43287571
#> CCR5    -2.8721780          -1.19817513                        0.14714636
#>        PopulationCD154+VbetaResponsive PopulationCD154+VbetaUnresponsive
#> B3GAT1                       1.6063715                        -0.5811225
#> BAX                         -0.4386521                        -0.5869687
#> BCL2                        -0.3862382                         1.3492612
#> CCL2                        -0.3052702                        -0.2626161
#> CCL3                         0.4259786                        -0.2617315
#> CCL4                         0.5515850                         0.3325102
#> CCL5                         0.6887183                         0.1555835
#> CCR2                        -0.5950754                         1.1230267
#> CCR4                         0.2698394                        -0.5880348
#> CCR5                         0.1345247                         0.6467222
#>        PopulationVbetaResponsive PopulationVbetaUnresponsive
#> B3GAT1                -1.2153975                 0.322296844
#> BAX                    0.1308370                 0.111068557
#> BCL2                   2.1094602                 1.928661427
#> CCL2                  -1.1332949                -2.305617623
#> CCL3                  -0.4638945                -2.238025106
#> CCL4                  -1.9149361                -1.918902719
#> CCL5                  -0.1764152                -0.305926046
#> CCR2                  -0.7973931                -0.150296597
#> CCR4                   0.1166578                 0.240411574
#> CCR5                  -0.3888459                -0.005858029

coef(zlmVbeta, 'C')

#>        (Intercept) Stim.ConditionUnstim PopulationCD154-VbetaUnresponsive
#> B3GAT1    18.19441           -1.7969063                       -0.55003066
#> BAX       17.47359           -0.6534516                       -0.08181114
#> BCL2      16.41403            1.3286670                       -0.86358539
#> CCL2      16.90490           -0.2937144                        3.11986916
#> CCL3      19.81370                   NA                       -0.55622889
#> CCL4      22.31976                   NA                       -5.32517963
#> CCL5      19.22605            0.4504418                       -0.68280704
#> CCR2      18.46733           -1.2269270                        0.82209897
#> CCR4      17.45862           -0.3702826                       -0.18436839
#> CCR5      15.77781            0.9461661                       -0.15253610
#>        PopulationCD154+VbetaResponsive PopulationCD154+VbetaUnresponsive
#> B3GAT1                              NA                                NA
#> BAX                         0.19649762                       -0.03911775
#> BCL2                        2.32150932                        0.94613777
#> CCL2                        7.04188443                        3.34328706
#> CCL3                        0.04812137                        3.32313947
#> CCL4                       -2.77190815                       -1.51398583
#> CCL5                        0.84757938                        0.64493103
#> CCR2                       -3.19304135                        0.98294260
#> CCR4                        0.57583680                        0.47465344
#> CCR5                        0.51211945                        1.66591584
#>        PopulationVbetaResponsive PopulationVbetaUnresponsive
#> B3GAT1                        NA                          NA
#> BAX                    0.1543621                   0.1850906
#> BCL2                   0.7362327                          NA
#> CCL2                          NA                          NA
#> CCL3                   0.5075641                          NA
#> CCL4                          NA                          NA
#> CCL5                  -0.7668751                  -0.4885039
#> CCR2                   0.1209953                  -1.1165337
#> CCR4                   0.3141728                  -0.1774801
#> CCR5                  -0.7502419                  -1.0153526

se.coef(zlmVbeta, 'C')

#>         
#> X1       (Intercept) Stim.ConditionUnstim PopulationCD154-VbetaUnresponsive
#>   B3GAT1   1.1195139            1.9390550                         1.9390550
#>   BAX      0.2834781            0.2625570                         0.3945849
#>   BCL2     0.5300655            1.2983899                         1.2983899
#>   CCL2     2.5224874            4.3690763                         3.5673358
#>   CCL3     2.2877129                   NA                         3.0263597
#>   CCL4     2.7577915                   NA                         4.7766350
#>   CCL5     0.9651997            1.0038683                         1.1357474
#>   CCR2     1.1047933            1.2757054                         1.3530899
#>   CCR4     0.4658746            0.4372988                         0.5966105
#>   CCR5     1.0524207            2.1048414                         1.2452410
#>         
#> X1       PopulationCD154+VbetaResponsive PopulationCD154+VbetaUnresponsive
#>   B3GAT1                              NA                                NA
#>   BAX                          0.3643741                         0.4482183
#>   BCL2                         1.2983899                         0.8655933
#>   CCL2                         3.5673358                         4.3690763
#>   CCL3                         2.6825812                         3.6171917
#>   CCL4                         3.2630629                         3.9001061
#>   CCL5                         1.0847908                         1.3068865
#>   CCR2                         1.5624136                         1.2757054
#>   CCR4                         0.5532457                         0.7298852
#>   CCR5                         1.2452410                         1.2889469
#>         
#> X1       PopulationVbetaResponsive PopulationVbetaUnresponsive
#>   B3GAT1                        NA                          NA
#>   BAX                    0.3620264                   0.3647365
#>   BCL2                   0.6940186                          NA
#>   CCL2                          NA                          NA
#>   CCL3                   3.6171917                          NA
#>   CCL4                          NA                          NA
#>   CCL5                   1.3267276                   1.2838356
#>   CCR2                   1.8596442                   1.5624136
#>   CCR4                   0.5968418                   0.5927323
#>   CCR5                   1.8228462                   1.3586693

#Test for a Population effect by dropping the whole term (a 5 degree of freedom test)
lrTest(zlmVbeta, 'Population')

#> Refitting on reduced model...

#> 
#> Done!

#> , , metric = lambda
#> 
#>         test.type
#> primerid      cont      disc    hurdle
#>   B3GAT1  0.000000  4.299060  4.299060
#>   BAX     1.072472 10.720978 11.793450
#>   BCL2    5.443797 19.740650 25.184447
#>   CCL2    4.103765  4.300998  8.404763
#>   CCL3    1.523584  7.074064  8.597648
#>   CCL4    1.516648  8.541585 10.058232
#>   CCL5    5.572644  5.419661 10.992304
#>   CCR2   12.540757  4.686256 17.227013
#>   CCR4    4.607460  8.733661 13.341121
#>   CCR5    7.136405  1.520790  8.657194
#> 
#> , , metric = df
#> 
#>         test.type
#> primerid cont disc hurdle
#>   B3GAT1    0    5      5
#>   BAX       5    5     10
#>   BCL2      4    5      9
#>   CCL2      3    5      8
#>   CCL3      4    5      9
#>   CCL4      3    5      8
#>   CCL5      5    5     10
#>   CCR2      5    5     10
#>   CCR4      5    5     10
#>   CCR5      5    5     10
#> 
#> , , metric = Pr(>Chisq)
#> 
#>         test.type
#> primerid       cont        disc     hurdle
#>   B3GAT1 1.00000000 0.507209744 0.50720974
#>   BAX    0.95651117 0.057201590 0.29911799
#>   BCL2   0.24471407 0.001397782 0.00277378
#>   CCL2   0.25047521 0.506941990 0.39496253
#>   CCL3   0.82245576 0.215190133 0.47521191
#>   CCL4   0.67843337 0.128806114 0.26096254
#>   CCL5   0.35004597 0.366835821 0.35811788
#>   CCR2   0.02808429 0.455356085 0.06948928
#>   CCR4   0.46563569 0.120170837 0.20521972
#>   CCR5   0.21069179 0.910658274 0.56491639
#> 
#> attr(,"test")
#> [1] "Population"

#Test only if the VbetaResponsive cells differ from the baseline group
lrTest(zlmVbeta, CoefficientHypothesis('PopulationVbetaResponsive'))

#> Refitting on reduced model...

#> 
#> Done!

#> , , metric = lambda
#> 
#>         test.type
#> primerid        cont       disc    hurdle
#>   B3GAT1 0.000000000 0.49967559 0.4996756
#>   BAX    0.187497834 0.05709950 0.2445973
#>   BCL2   0.000000000 7.50139598 7.5013960
#>   CCL2   0.000000000 1.51305236 1.5130524
#>   CCL3   0.022203431 0.31780369 0.3400071
#>   CCL4   0.000000000 2.44968166 2.4496817
#>   CCL5   0.357112181 0.06828354 0.4253957
#>   CCR2   0.006224572 0.58399389 0.5902185
#>   CCR4   0.284055573 0.07376621 0.3578218
#>   CCR5   0.210197743 0.11201241 0.3222102
#> 
#> , , metric = df
#> 
#>         test.type
#> primerid cont disc hurdle
#>   B3GAT1    0    1      1
#>   BAX       1    1      2
#>   BCL2      0    1      1
#>   CCL2      0    1      1
#>   CCL3      1    1      2
#>   CCL4      0    1      1
#>   CCL5      1    1      2
#>   CCR2      1    1      2
#>   CCR4      1    1      2
#>   CCR5      1    1      2
#> 
#> , , metric = Pr(>Chisq)
#> 
#>         test.type
#> primerid      cont        disc      hurdle
#>   B3GAT1 1.0000000 0.479642699 0.479642699
#>   BAX    0.6650074 0.811140466 0.884884044
#>   BCL2   1.0000000 0.006165119 0.006165119
#>   CCL2   1.0000000 0.218673913 0.218673913
#>   CCL3   0.8815472 0.572930548 0.843661814
#>   CCL4   1.0000000 0.117548705 0.117548705
#>   CCL5   0.5501144 0.793852536 0.808400350
#>   CCR2   0.9371154 0.444751088 0.744450266
#>   CCR4   0.5940553 0.785930097 0.836180408
#>   CCR5   0.6466125 0.737864563 0.851202624
#> 
#> attr(,"test")
#> [1] "PopulationVbetaResponsive"

# Test if there is a difference between CD154+/Unresponsive and CD154-/Unresponsive.
# Note that because we parse the expression
# the columns must be enclosed in backquotes
# to protect the \quote{+} and \quote{-} characters.
lrTest(zlmVbeta, Hypothesis('`PopulationCD154+VbetaUnresponsive` -
        `PopulationCD154-VbetaUnresponsive`'))

#> Warning: Some levels contain symbols.  Be careful to escape these names with backticks ('`') when specifying contrasts.

#> Refitting on reduced model...

#> 
#> Done!

#> , , metric = lambda
#> 
#>         test.type
#> primerid        cont        disc     hurdle
#>   B3GAT1 0.000000000  0.53957307  0.5395731
#>   BAX    0.009603320  0.41964445  0.4292478
#>   BCL2   2.004979393  2.89187997  4.8968594
#>   CCL2   0.003581828 -0.15725169 -0.1536699
#>   CCL3   1.389919527 -0.60229580  0.7876237
#>   CCL4   0.695474388  1.22512019  1.9205946
#>   CCL5   1.643474587  0.02224792  1.6657225
#>   CCR2   0.037354159  1.49555138  1.5329055
#>   CCR4   0.977322744  0.14461070  1.1219334
#>   CCR5   3.776168456  0.50471809  4.2808865
#> 
#> , , metric = df
#> 
#>         test.type
#> primerid cont disc hurdle
#>   B3GAT1    0    1      1
#>   BAX       1    1      2
#>   BCL2      1    1      2
#>   CCL2      1    1      2
#>   CCL3      1    1      2
#>   CCL4      1    1      2
#>   CCL5      1    1      2
#>   CCR2      1    1      2
#>   CCR4      1    1      2
#>   CCR5      1    1      2
#> 
#> , , metric = Pr(>Chisq)
#> 
#>         test.type
#> primerid       cont       disc    hurdle
#>   B3GAT1 1.00000000 0.46260972 0.4626097
#>   BAX    0.92193505 0.51711451 0.8068449
#>   BCL2   0.15678342 0.08902699 0.0864292
#>   CCL2   0.95227640 1.00000000 1.0000000
#>   CCL3   0.23841869 1.00000000 0.6744809
#>   CCL4   0.40430856 0.26835815 0.3827791
#>   CCL5   0.19984941 0.88142947 0.4348034
#>   CCR2   0.84674577 0.22135713 0.4646584
#>   CCR4   0.32286067 0.70373971 0.5706571
#>   CCR5   0.05198758 0.47743434 0.1176027
#> 
#> attr(,"test")
#> [1] "Contrast Matrix"

waldTest(zlmVbeta, Hypothesis('`PopulationCD154+VbetaUnresponsive` -
        `PopulationCD154-VbetaUnresponsive`'))

#> Warning: Some levels contain symbols.  Be careful to escape these names with backticks ('`') when specifying contrasts.

#> , , metric = lambda
#> 
#>         test.type
#> primerid        cont         disc      hurdle
#>   B3GAT1          NA 0.4246764991          NA
#>   BAX    0.009305455 0.4499539395 0.459259394
#>   BCL2   1.748463371 2.4975649351 4.246028306
#>   CCL2   0.002614910 0.0007340968 0.003349006
#>   CCL3   1.278016527 0.0007253470 1.278741874
#>   CCL4   0.636616606 1.2610150758 1.897631682
#>   CCL5   1.553679693 0.0235791990 1.577258892
#>   CCR2   0.025434721 1.4461225520 1.471557273
#>   CCR4   0.955370495 0.1340900313 1.089460527
#>   CCR5   3.317281808 0.5214642628 3.838746071
#> 
#> , , metric = df
#> 
#>         test.type
#> primerid cont disc hurdle
#>   B3GAT1    1    1      2
#>   BAX       1    1      2
#>   BCL2      1    1      2
#>   CCL2      1    1      2
#>   CCL3      1    1      2
#>   CCL4      1    1      2
#>   CCL5      1    1      2
#>   CCR2      1    1      2
#>   CCR4      1    1      2
#>   CCR5      1    1      2
#> 
#> , , metric = Pr(>Chisq)
#> 
#>         test.type
#> primerid       cont      disc    hurdle
#>   B3GAT1         NA 0.5146127        NA
#>   BAX    0.92315144 0.5023568 0.7948279
#>   BCL2   0.18607002 0.1140225 0.1196704
#>   CCL2   0.95921700 0.9783846 0.9983269
#>   CCL3   0.25826815 0.9785138 0.5276242
#>   CCL4   0.42493864 0.2614590 0.3871993
#>   CCL5   0.21259304 0.8779605 0.4544672
#>   CCR2   0.87328861 0.2291511 0.4791322
#>   CCR4   0.32835604 0.7142285 0.5799982
#>   CCR5   0.06855509 0.4702177 0.1466989
#>