Deprecated cell/feature data accessors/mutators

These functions are now all deprecated and will be removed in a future release.

cData(sc)

cData(sc) <- value

# S4 method for SingleCellAssay
cData(sc)

# S4 method for SingleCellAssay
cData(sc) <- value

# S4 method for SingleCellAssay,SingleCellAssay
combine(x, y, ...)

# S4 method for SingleCellAssay,ANY
combine(x, y, ...)

Arguments

sc

An object with cellData

value

replacement value

x

SingleCellAssay

y

SingleCellAssay

...

SingleCellAssay

Value

DataFrame or modifies the SingleCellAssay object in place

Details

cData(sc): Return the cellData data.frame.

cData(sc)<-value: Replace the cellData with value, which can be either an AnnotatedDataFrame or data.frame. The replacement is checked that it has mandatory fields defined by its class.

combine(x, y, ...): Concatenate two experiments along rows/columns

Replacement Functions

You should transition to use the following replacements:

cData

colData

fData

mcols

exprs

assay

combine

cbind2 or rbind2

See also

exprs

Examples

data(vbetaFA)
stopifnot(all.equal(hushWarning(cData(vbetaFA), 'deprecated'),colData(vbetaFA)))
stopifnot(all.equal(hushWarning(fData(vbetaFA), 'deprecated'), mcols(vbetaFA)))
stopifnot(all.equal(hushWarning(exprs(vbetaFA), 'deprecated'), t(assay(vbetaFA))))