Convert a MASTClassic SingleCellAssay

Convert a SingleCellAssay object created with the MASTClassic package to an object recognized by the new MAST package

convertMASTClassicToSingleCellAssay(object = NULL)

Arguments

object

of class SingleCellAssay created by MASTClassic

Value

A MAST SingleCellAssay object.

Details

The function will extract the relevant information from the attributes of the old object and construct a new SingleCellAssay that is recognized by MAST. This function checks that the object is a MASTClassic SingleCellAssay object. It will stop if it is not a SingleCellAssay, return a converted SingleCellAssay if object was created by MASTClassic, and return the original object if the object is already compatible.

Note

Type checking for old object is not performed.

Examples

data(vbetaFA)
convertMASTClassicToSingleCellAssay(vbetaFA)
#> object does not appear to be created by MASTClassic. Returning original.
#> class: FluidigmAssay 
#> dim: 75 456 
#> metadata(0):
#> assays(1): Et
#> rownames(75): B3GAT1 BAX ... TNFRSF9 TNFSF10
#> rowData names(2): Gene primerid
#> colnames(456): Sub01 1 A01 Sub01 1 A02 ... Sub02 3 H10 Sub02 3 H11
#> colData names(9): Number.of.Cells Population ... Time wellKey
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):