Filter low-expressing genes

Filter out genes that have less than some percent threshold expression across all libraries

filterLowExpressedGenes(assay, threshold = 0.1)

Arguments

assay

a SingleCellAssay object

threshold

a numeric between 0, and 1, specifying the threshold frequency below which genes will be filtered out

Value

SingleCellAssay

Examples

data(vbetaFA)
filterLowExpressedGenes(vbetaFA)
#> Filtered out 72.00 percent of genes at 10.00 percent threshold
#> class: FluidigmAssay 
#> dim: 54 456 
#> metadata(0):
#> assays(1): Et
#> rownames(54): BAX BCL2 ... TNFRSF9 TNFSF10
#> rowData names(2): Gene primerid
#> colnames(456): Sub01 1 A01 Sub01 1 A02 ... Sub02 3 H10 Sub02 3 H11
#> colData names(9): Number.of.Cells Population ... Time wellKey
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):