Plot a likelihood ratio test object

Constructs a forest-like plot of signed log10 p-values, possibly adjusted for multiple comparisons adjust can be one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

plotlrt(lr, adjust = "fdr", thres = 0.1, trunc = 1e-06, groups = NULL)

Arguments

lr

output from lrtest, with returnall=FALSE

adjust

character, passed along to p.adjust, see below

thres

numeric genes with adjusted pvalues above this value are not depicted

trunc

numeric p values below this value are truncated at this value

groups

character grouping value. If provided, must match groups argument passed to lrtest. Plots done separately for each group.

Value

Constructs a dotplot

Author

andrew