Reads a Fluidigm Biomark (c. 2011) raw data file (or set of files)

This function reads a raw Fluidigm Biomark data file or set of files and constructs a SingleCellAssay (or FluigidmAssay) object. This was written c. 2011 and has not been tested lately. The Biomark format may have changed.

read.fluidigm(
  files = NULL,
  metadata = NULL,
  header.size = 2,
  skip = 8,
  cycle.threshold = 40,
  metadataColClasses = NULL,
  meta.key = NULL,
  idvars = NULL,
  splitby = NULL,
  unique.well.id = "Chamber.ID",
  raw = TRUE,
  assay = NULL,
  geneid = "Assay.Name",
  sample = NULL,
  well = "Well",
  measurement = "X40.Ct",
  measurement.processed = "Ct",
  ncells = "SampleRConc"
)

Arguments

files

A character vector of files to read.

metadata

A character path and filename of a CSV file containing additional metadata about the samples

header.size

A numeric indicating the number of lines in the header (default 2)

skip

numeric how many lines to skip before reading (default 8)

cycle.threshold

The maximum number of PCR cycles performed (default 40) numeric

metadataColClasses

Optional character vector giving the column classes of the metadata file. See read.table.

meta.key

Optional character vector that identifies the key column between the metadata and the fluidigm data

idvars

Optional character vector that defines the set of columns uniquely identifying a well (unique cell, gene, and condition).

splitby

Optional character that defines the column / variable used to split the resulting data into a list of SingleCellAssay, such that unique levels of splitby each fall into their own SingleCellAssay. Ususally the experimental unit subjected to different treatments.

unique.well.id

The column that uniquely identifies a sample well in the data. Default is "Chamber.ID".

raw

logical flag indicating this is raw data coming off the instrument. Thus we make some assumptions about the column names that are present.

assay

character name of a column that uniquely identifies an Assay (i.e. gene). Default is NULL

geneid

character names of the column that identifies a gene. Default is "Assay.Name"

sample

character name of a column that uniquely identifies a sample

well

character name of a column that uniquely identifies a well. Default "Well".

measurement

character name of the column that holds the measurement. Default "X40.Ct".

measurement.processed

character one of "Ct","40-Ct", or "et". If not "Ct", the measurement will be transformed.

ncells

The column with the number of cells in this well.

Value

list of SingleCellAssay holding the data.

Author

Greg Finak