Summarize model features from a ZlmFit object

Returns a data.table with a special print method that shows the top 2 most significant genes by contrast. This data.table contains columns:

primerid

the gene

component

C=continuous, D=discrete, logFC=log fold change, S=combined using Stouffer's method, H=combined using hurdle method

contrast

the coefficient/contrast of interest

ci.hi

upper bound of confidence interval

ci.lo

lower bound of confidence interval

coef

point estimate

z

z score (coefficient divided by standard error of coefficient)

Pr(>Chisq)

likelihood ratio test p-value (only if doLRT=TRUE)

Some of these columns will contain NAs if they are not applicable for a particular component or contrast.

# S4 method for ZlmFit
summary(object, logFC = TRUE, doLRT = FALSE,
  level = 0.95, ...)

Arguments

object
A ZlmFit object

logFC
If TRUE, calculate log-fold changes, or output from a call to getLogFC.

doLRT
if TRUE, calculate lrTests on each coefficient, or a character vector of such coefficients to consider.

level
what level of confidence coefficient to return. Defaults to 95 percent.

...
ignored

Value

data.table

See also

Examples

data(vbetaFA)
z <- zlm(~Stim.Condition, vbetaFA[1:5,])

#> 
#> Done!

zs <- summary(z)

#> Combining coefficients and standard errors

#> Calculating log-fold changes

names(zs)

#> [1] "datatable"

print(zs)

#> Fitted zlm with top 2 genes per contrast:
#> ( log fold change Z-score )
#>  primerid Stim.ConditionUnstim
#>  BCL2       -2.3*             
#>  CCL3       -3.8*             

##Select `datatable` copmonent to get normal print method
zs$datatable

#>     primerid component             contrast       ci.hi      ci.lo       coef
#>  1:     CCL2         D          (Intercept) -2.44236757 -3.3620226 -2.9021951
#>  2:     CCL3         D          (Intercept) -1.58916186 -2.1962433 -1.8927026
#>  3:     BCL2         D          (Intercept) -1.56579681 -2.1670394 -1.8664181
#>  4:   B3GAT1         D          (Intercept) -2.69053309 -3.7525507 -3.2215419
#>  5:      BAX         D          (Intercept) -0.33128091 -0.7564909 -0.5438859
#>  6:     CCL2         S          (Intercept)          NA         NA         NA
#>  7:     CCL3         S          (Intercept)          NA         NA         NA
#>  8:     CCL2         C          (Intercept) 22.37163738 19.2284654 20.8000514
#>  9:     CCL3         C          (Intercept) 23.15685816 20.9696836 22.0632709
#> 10:   B3GAT1         S          (Intercept)          NA         NA         NA
#> 11:     BCL2         S          (Intercept)          NA         NA         NA
#> 12:      BAX         S          (Intercept)          NA         NA         NA
#> 13:   B3GAT1         C          (Intercept) 18.66804935 17.6920962 18.1800728
#> 14:     BCL2         C          (Intercept) 19.41604938 18.4701211 18.9430853
#> 15:      BAX         C          (Intercept) 19.31259333 18.4934991 18.9030462
#> 16:     CCL3     logFC Stim.ConditionUnstim -1.03709089 -3.2491890 -2.1431399
#> 17:     CCL3         D Stim.ConditionUnstim -0.28878236 -2.5303707 -1.4095765
#> 18:      BAX         C Stim.ConditionUnstim -0.21303964 -1.9052758 -1.0591577
#> 19:     BCL2     logFC Stim.ConditionUnstim -0.17786584 -2.4494721 -1.3136690
#> 20:     CCL3         S Stim.ConditionUnstim          NA         NA         NA
#> 21:     BCL2         S Stim.ConditionUnstim          NA         NA         NA
#> 22:     BCL2         D Stim.ConditionUnstim  0.09775406 -1.6221758 -0.7622108
#> 23:     BCL2         C Stim.ConditionUnstim  0.66939860 -2.1945414 -0.7625714
#> 24:   B3GAT1         C Stim.ConditionUnstim  0.57039007 -1.7528431 -0.5912265
#> 25:   B3GAT1         S Stim.ConditionUnstim          NA         NA         NA
#> 26:     CCL2         C Stim.ConditionUnstim  2.22026031 -5.3168014 -1.5482705
#> 27:     CCL2         S Stim.ConditionUnstim          NA         NA         NA
#> 28:      BAX         S Stim.ConditionUnstim          NA         NA         NA
#> 29:     CCL2     logFC Stim.ConditionUnstim  0.70894380 -1.1676533 -0.2293547
#> 30:     CCL3         C Stim.ConditionUnstim  3.45685041 -5.5611013 -1.0521254
#> 31:     CCL2         D Stim.ConditionUnstim  0.90444296 -1.2421537 -0.1688554
#> 32:   B3GAT1     logFC Stim.ConditionUnstim  0.60696199 -0.8305510 -0.1117945
#> 33:   B3GAT1         D Stim.ConditionUnstim  1.07670176 -1.3704909 -0.1468946
#> 34:      BAX     logFC Stim.ConditionUnstim  3.15942619 -0.9870469  1.0861896
#> 35:      BAX         D Stim.ConditionUnstim  0.80556556 -0.1194848  0.3430404
#>     primerid component             contrast       ci.hi      ci.lo       coef
#>               z
#>  1: -12.3702861
#>  2: -12.2211896
#>  3: -12.1685068
#>  4: -11.8907750
#>  5:  -5.0139783
#>  6:   9.5954215
#>  7:  19.3191106
#>  8:  25.9402614
#>  9:  39.5425379
#> 10:  43.2252343
#> 11:  46.9035705
#> 12:  60.4223239
#> 13:  73.0204876
#> 14:  78.5001723
#> 15:  90.4640483
#> 16:  -3.7977313
#> 17:  -2.4649658
#> 18:  -2.4534531
#> 19:  -2.2668927
#> 20:  -2.0663811
#> 21:  -1.9664048
#> 22:  -1.7371707
#> 23:  -1.0437456
#> 24:  -0.9975604
#> 25:  -0.8717615
#> 26:  -0.8052354
#> 27:  -0.7874231
#> 28:  -0.7069721
#> 29:  -0.4790874
#> 30:  -0.4573384
#> 31:  -0.3083490
#> 32:  -0.3048504
#> 33:  -0.2352966
#> 34:   1.0268450
#> 35:   1.4536435
#>               z