Summarize the effect of thresholding

Returns the proportion of (putative) expression, the variance of expressed cells, and -log10 shapiro-wilk tests for normality on the expressed cells

# S3 method for thresholdSCRNACountMatrix
summary(object, ...)

# S3 method for summaryThresholdSCRNA
print(x, ...)

Arguments

object

a thresholdSCRNACountMatrix

...

currently ignored

x

a summaryThresholdSCRNA object, ie output from summary.thresholdSCRNACountMatrix

Value

a list of statistics on the original data, and thresholded data

Methods (by generic)

print: prints five-number distillation of the statistics and invisibly returns the table used to generate the summary