Likelihood Ratio Tests for SingleCellAssays

Tests for a change in ET binomial proportion or mean of positive ET Likelihood Ratio Test for SingleCellAssay objects

LRT(sca, comparison, ...)

# S4 method for SingleCellAssay,character
LRT(sca, comparison, referent = NULL, groups = NULL, returnall = FALSE)

Arguments

sca

A SingleCellAssay class object

comparison

A character specifying the factor for comparison

...

ignored

referent

A character specifying the reference level of comparison.

groups

A optional character specifying a variable on which to stratify the test. For each level of groups, there will be a separate likelihood ratio test.

returnall

A logical specifying if additional rows should be returned with information about the different components of the test.

Value

data.frame

Details

Combined Likelihood ratio test (binomial and normal) for SingleCellAssay and derived objects. This function is deprecated, please use lrTest instead.

See also

Examples

data(vbetaFA)
LRT(vbetaFA, 'Stim.Condition', 'Unstim')
#>    Stim.Condition test.type primerid     lrstat direction      p.value
#> 1       Stim(SEB)      comb   B3GAT1  1.5622613         1 4.578880e-01
#> 2       Stim(SEB)      comb      BAX  8.0652869         1 1.772741e-02
#> 3       Stim(SEB)      comb     BCL2  4.7450627         1 9.324439e-02
#> 4       Stim(SEB)      comb     CCL2  0.7561235         1 6.851882e-01
#> 5       Stim(SEB)      comb     CCL3  9.2095911         1 1.000375e-02
#> 6       Stim(SEB)      comb     CCL4 13.8735075         1 9.714180e-04
#> 7       Stim(SEB)      comb     CCL5  1.7575976         1 4.152814e-01
#> 8       Stim(SEB)      comb     CCR2  2.8985299         1 2.347428e-01
#> 9       Stim(SEB)      comb     CCR4  3.3223886         1 1.899120e-01
#> 10      Stim(SEB)      comb     CCR5  4.6587323         1 9.735744e-02
#> 11      Stim(SEB)      comb     CCR7 39.1243255        -1 3.193454e-09
#> 12      Stim(SEB)      comb    CD109 11.1843027         1 3.727001e-03
#> 13      Stim(SEB)      comb     CD27 39.9679896        -1 2.094408e-09
#> 14      Stim(SEB)      comb     CD28 24.7306020        -1 4.264008e-06
#> 15      Stim(SEB)      comb     CD38  4.6791199         1 9.637004e-02
#> 16      Stim(SEB)      comb     CD3d  9.2052879         1 1.002529e-02
#>  [ reached 'max' / getOption("max.print") -- omitted 59 rows ]