Likelihood Ratio Tests for SingleCellAssays

Tests for a change in ET binomial proportion or mean of positive ET Likelihood Ratio Test for SingleCellAssay objects

LRT(sca, comparison, ...)

# S4 method for SingleCellAssay,character
LRT(sca, comparison,
  referent = NULL, groups = NULL, returnall = FALSE)

Arguments

sca
A SingleCellAssay class object

comparison
A character specifying the factor for comparison

...
ignored

referent
A character specifying the reference level of comparison.

groups
A optional character specifying a variable on which to stratify the test. For each level of groups, there will be a separate likelihood ratio test.

returnall
A logical specifying if additional rows should be returned with information about the different components of the test.

Value

data.frame

Details

Combined Likelihood ratio test (binomial and normal) for SingleCellAssay and derived objects. This function is deprecated, please use lrTest instead.

See also

Examples

data(vbetaFA)
LRT(vbetaFA, 'Stim.Condition', 'Unstim')

#>    Stim.Condition test.type primerid     lrstat direction      p.value
#> 1       Stim(SEB)      comb   B3GAT1  1.5622613         1 4.578880e-01
#> 2       Stim(SEB)      comb      BAX  8.0652869         1 1.772741e-02
#> 3       Stim(SEB)      comb     BCL2  4.7450627         1 9.324439e-02
#> 4       Stim(SEB)      comb     CCL2  0.7561235         1 6.851882e-01
#> 5       Stim(SEB)      comb     CCL3  9.2095911         1 1.000375e-02
#> 6       Stim(SEB)      comb     CCL4 13.8735075         1 9.714180e-04
#> 7       Stim(SEB)      comb     CCL5  1.7575976         1 4.152814e-01
#> 8       Stim(SEB)      comb     CCR2  2.8985299         1 2.347428e-01
#> 9       Stim(SEB)      comb     CCR4  3.3223886         1 1.899120e-01
#> 10      Stim(SEB)      comb     CCR5  4.6587323         1 9.735744e-02
#> 11      Stim(SEB)      comb     CCR7 39.1243255        -1 3.193454e-09
#> 12      Stim(SEB)      comb    CD109 11.1843027         1 3.727001e-03
#> 13      Stim(SEB)      comb     CD27 39.9679896        -1 2.094408e-09
#> 14      Stim(SEB)      comb     CD28 24.7306020        -1 4.264008e-06
#> 15      Stim(SEB)      comb     CD38  4.6791199         1 9.637004e-02
#> 16      Stim(SEB)      comb     CD3d  9.2052879         1 1.002529e-02
#> 17      Stim(SEB)      comb     CD3g 12.2316534        -1 2.207650e-03
#> 18      Stim(SEB)      comb      CD4 20.3969774         1 3.722654e-05
#> 19      Stim(SEB)      comb   CD40LG 47.1325025         1 5.825150e-11
#> 20      Stim(SEB)      comb     CD45  2.9793280         1 2.254484e-01
#> 21      Stim(SEB)      comb     CD48 22.0357926        -1 1.640546e-05
#> 22      Stim(SEB)      comb      CD5 35.8896419        -1 1.609397e-08
#> 23      Stim(SEB)      comb     CD69 59.6283282         1 1.126868e-13
#> 24      Stim(SEB)      comb     CD8a  1.9720800         1 3.730511e-01
#> 25      Stim(SEB)      comb     CSF1 14.1877582         1 8.301708e-04
#> 26      Stim(SEB)      comb     CSF2 12.4373631         1 1.991870e-03
#> 27      Stim(SEB)      comb    CTLA4  5.4613792         1 6.517433e-02
#> 28      Stim(SEB)      comb     CTSD 32.7168485        -1 7.863694e-08
#> 29      Stim(SEB)      comb   CXCL13  1.7895695         1 4.086956e-01
#> 30      Stim(SEB)      comb    CXCR4 24.7689912        -1 4.182942e-06
#> 31      Stim(SEB)      comb     DPP4 11.6187037         1 2.999373e-03
#> 32      Stim(SEB)      comb   ENTPD1  1.7560858         1 4.155955e-01
#> 33      Stim(SEB)      comb    EOMES  2.2173355         1 3.299983e-01
#> 34      Stim(SEB)      comb      FAS  6.2726399         1 4.344237e-02
#> 35      Stim(SEB)      comb    FASLG 12.6641628         1 1.778329e-03
#> 36      Stim(SEB)      comb    FOXP3  8.8656802         1 1.188070e-02
#> 37      Stim(SEB)      comb    GAPDH 41.8965729         1 7.984998e-10
#> 38      Stim(SEB)      comb    GATA3  7.2460169         1 2.670222e-02
#> 39      Stim(SEB)      comb     GZMA  0.4497203        -1 7.986279e-01
#> 40      Stim(SEB)      comb   HLADRA  4.5857404         1 1.009762e-01
#> 41      Stim(SEB)      comb     ICOS  1.6924939         1 4.290221e-01
#> 42      Stim(SEB)      comb     IFNg 19.0305726         1 7.371632e-05
#> 43      Stim(SEB)      comb     IL10 15.2728700         1 4.825457e-04
#> 44      Stim(SEB)      comb IL12RbII  0.9040240         1 6.363465e-01
#> 45      Stim(SEB)      comb     IL13  6.8040022         1 3.330655e-02
#> 46      Stim(SEB)      comb     IL16 80.6806088        -1 3.022932e-18
#> 47      Stim(SEB)      comb      IL2 21.4202359         1 2.231798e-05
#> 48      Stim(SEB)      comb     IL22 10.0688053         1 6.510086e-03
#> 49      Stim(SEB)      comb    IL2Ra 16.5345877         1 2.567792e-04
#> 50      Stim(SEB)      comb    IL2Rg 27.3758616        -1 1.136075e-06
#> 51      Stim(SEB)      comb      IL3  4.0516709         1 1.318836e-01
#> 52      Stim(SEB)      comb      IL4  4.9644775         1 8.355595e-02
#> 53      Stim(SEB)      comb     IL6R 45.7493495        -1 1.163202e-10
#> 54      Stim(SEB)      comb    IL6ST 13.8079441        -1 1.003790e-03
#> 55      Stim(SEB)      comb     IL7R 12.6296402        -1 1.809291e-03
#> 56      Stim(SEB)      comb      IL8  9.7769904         1 7.532749e-03
#> 57      Stim(SEB)      comb      LIF 16.0247377         1 3.313389e-04
#> 58      Stim(SEB)      comb    MKI67  1.3405178         1 5.115761e-01
#> 59      Stim(SEB)      comb     MMP9 11.2754562        -1 3.560949e-03
#> 60      Stim(SEB)      comb     NT5E  2.6568509         1 2.648940e-01
#> 61      Stim(SEB)      comb     PCNA  3.2187139         1 2.000162e-01
#> 62      Stim(SEB)      comb   PECAM1  1.7650930         1 4.137280e-01
#> 63      Stim(SEB)      comb    PRDM1 10.3503530         1 5.655219e-03
#> 64      Stim(SEB)      comb     PRF1  2.5049682         1 2.857940e-01
#> 65      Stim(SEB)      comb     RORC  4.0621171         1 1.311966e-01
#> 66      Stim(SEB)      comb   SLAMF1  9.8238721         1 7.358229e-03
#> 67      Stim(SEB)      comb   SLAMF5  2.4285892         1 2.969194e-01
#> 68      Stim(SEB)      comb   SLAMF7  4.9270545         1 8.513413e-02
#> 69      Stim(SEB)      comb    TBX21  3.5895831         1 1.661621e-01
#> 70      Stim(SEB)      comb    TGFB1 17.1380553         1 1.898972e-04
#> 71      Stim(SEB)      comb     TIA1  5.2156846         1 7.369338e-02
#> 72      Stim(SEB)      comb    TIMP1  4.6041022         1 1.000534e-01
#> 73      Stim(SEB)      comb      TNF 32.3906762         1 9.256654e-08
#> 74      Stim(SEB)      comb  TNFRSF9 23.7068304         1 7.114218e-06
#> 75      Stim(SEB)      comb  TNFSF10  5.7445662        -1 5.656962e-02