Summarize model features from a

Returns a data.table with a special print method that shows the top 2 most significant genes by contrast. This data.table contains columns:

primerid

the gene

component

C=continuous, D=discrete, logFC=log fold change, S=combined using Stouffer's method, H=combined using hurdle method

contrast

the coefficient/contrast of interest

ci.hi

upper bound of confidence interval

ci.lo

lower bound of confidence interval

coef

point estimate

z

z score (coefficient divided by standard error of coefficient)

Pr(>Chisq)

likelihood ratio test p-value (only if doLRT=TRUE)

Some of these columns will contain NAs if they are not applicable for a particular component or contrast.

# S4 method for ZlmFit
summary(object, logFC = TRUE, doLRT = FALSE,
  level = 0.95, ...)

Arguments

object

A ZlmFit object

logFC

If TRUE, calculate log-fold changes, or output from a call to getLogFC.

doLRT

if TRUE, calculate lrTests on each coefficient, or a character vector of such coefficients to consider.

level

what level of confidence coefficient to return. Defaults to 95 percent.

...

ignored

Value

data.table

See also

print.summaryZlmFit

Examples

data(vbetaFA)
z <- zlm(~Stim.Condition, vbetaFA[1:5,])

#> 
#> Done!

zs <- summary(z)

#> Combining coefficients and standard errors

#> Calculating log-fold changes

names(zs)

#> [1] "datatable"

print(zs)

#> Fitted zlm with top 2 genes per contrast:
#> ( log fold change Z-score )
#>  primerid Stim.ConditionUnstim
#>  BCL2       -2.3*             
#>  CCL3       -3.8*             

##Select `datatable` copmonent to get normal print method
zs$datatable

#>     primerid component             contrast       ci.hi      ci.lo       coef
#>  1:     CCL2         D          (Intercept) -2.44236757 -3.3620226 -2.9021951
#>  2:     CCL3         D          (Intercept) -1.58916186 -2.1962433 -1.8927026
#>  3:     BCL2         D          (Intercept) -1.56579681 -2.1670394 -1.8664181
#>  4:   B3GAT1         D          (Intercept) -2.69053309 -3.7525507 -3.2215419
#>  5:      BAX         D          (Intercept) -0.33128091 -0.7564909 -0.5438859
#>  6:     CCL2         S          (Intercept)          NA         NA         NA
#>  7:     CCL3         S          (Intercept)          NA         NA         NA
#>  8:     CCL2         C          (Intercept) 22.37163738 19.2284654 20.8000514
#>  9:     CCL3         C          (Intercept) 23.15685816 20.9696836 22.0632709
#> 10:   B3GAT1         S          (Intercept)          NA         NA         NA
#> 11:     BCL2         S          (Intercept)          NA         NA         NA
#> 12:      BAX         S          (Intercept)          NA         NA         NA
#> 13:   B3GAT1         C          (Intercept) 18.66804935 17.6920962 18.1800728
#> 14:     BCL2         C          (Intercept) 19.41604938 18.4701211 18.9430853
#>               z
#>  1: -12.3702861
#>  2: -12.2211896
#>  3: -12.1685068
#>  4: -11.8907750
#>  5:  -5.0139783
#>  6:   9.5954215
#>  7:  19.3191106
#>  8:  25.9402614
#>  9:  39.5425379
#> 10:  43.2252343
#> 11:  46.9035705
#> 12:  60.4223239
#> 13:  73.0204876
#> 14:  78.5001723
#>  [ reached getOption("max.print") -- omitted 22 rows ]